algorithms/java/src/main/java/org/jessenpan/leetcode/array/S334IncreasingTripletSubsequence.java
package org.jessenpan.leetcode.array;

/**
 * @author jessenpan
 * tag:array
 */
public class S334IncreasingTripletSubsequence {

    public boolean increasingTriplet(int[] nums) {

        int least = Integer.MAX_VALUE, mid = Integer.MAX_VALUE;
        for (int num : nums) {
            if (num <= least) {
                least = num;
            } else if (num <= mid) {
                mid = num;
            } else {
                return true;
            }
        }

        return false;
    }

}
